refactor: LotteryResult 에 대한 추가 테스트가 필요하여 새로운 테스트 클래스로 이동
