+++ src/test/java/step2/WinningNumberTest.java
@@ -9,10 +9,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Supplier;
 import java.util.stream.Stream;
 
 import static java.util.Collections.singletonList;
@@ -21,8 +18,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static step2.LotteryAgentTest.LotteryTickets;
 import static step2.LotteryNumberTest.LotteryNumber;
 import static step2.PlayslipTest.Playslip.NUMBER_POOL;
-import static step2.WinningNumberTest.LotteryResult.Matched;
-import static step2.WinningNumberTest.LotteryResult.Matched.*;
+import static step2.LotteryResultTest.LotteryResult.Matched;
+import static step2.LotteryResultTest.LotteryResult.Matched.*;
 
 
 public class WinningNumberTest {
@@ -56,7 +53,7 @@ public class WinningNumberTest {
         Integer[] numbers = Arrays.stream(stringNumbers.split(",")) //
                 .map(Integer::valueOf) //
                 .toArray(Integer[]::new);
-        LotteryResult lotteryResult = winningNumber.match(makeLotteryTickets(numbers));
+        LotteryResultTest.LotteryResult lotteryResult = winningNumber.match(makeLotteryTickets(numbers));
 
         assertThat(lotteryResult.getMatchResult(Matched.miss)).isEqualTo(miss);
     }
@@ -66,7 +63,7 @@ public class WinningNumberTest {
     @MethodSource("matchingCountProvider")
     void matched(LotteryNumber lotteryNumber, Matched matched) {
         //@formatter:off
-        LotteryResult lotteryResult = winningNumber.match(new LotteryTickets(
+        LotteryResultTest.LotteryResult lotteryResult = winningNumber.match(new LotteryTickets(
                 Arrays.asList(
                         lotteryNumber,
                         lotteryNumber,
@@ -99,12 +96,12 @@ public class WinningNumberTest {
             this.numberSelection = numberSelection;
         }
 
-        public LotteryResult match(LotteryTickets tickets) {
+        public LotteryResultTest.LotteryResult match(LotteryTickets tickets) {
             if (winningNumber == null) {
                 throw new IllegalStateException();
             }
 
-            LotteryResult lotteryResult = new LotteryResult();
+            LotteryResultTest.LotteryResult lotteryResult = new LotteryResultTest.LotteryResult();
             for (LotteryNumber ticketNumber : tickets.getNumbers()) {
                 lotteryResult.add(ticketNumber.getMatched(winningNumber));
             }
@@ -135,37 +132,4 @@ public class WinningNumberTest {
         }
     }
 
-    static class LotteryResult {
-        enum Matched {
-            miss(0), three(3), four(4), five(5), six(6);
-
-            private final int matchingCount;
-
-            Matched(int matchingCount) {
-                this.matchingCount = matchingCount;
-            }
-
-            public static Matched valueBy(Supplier<Integer> matched) {
-                return Arrays.stream(values()) //
-                        .filter(value -> value.matchingCount == matched.get()) //
-                        .findFirst().orElseThrow(IllegalArgumentException::new);
-            }
-        }
-
-        private final Map<Matched, Integer> result = new HashMap<>();
-
-        public void add(int matched) {
-            Matched matchedEnum = Matched.valueBy(() -> {
-                if (matched < 3) {
-                    return 0;
-                } return matched;
-            });
-
-            result.compute(matchedEnum, (key, value) -> value == null ? 1 : value + 1);
-        }
-
-        public int getMatchResult(Matched matched) {
-            return result.getOrDefault(matched, 0);
-        }
-    }
 }
